==> 03_chatglm_feats.ipynb <==
import pandas as pd
import numpy as np

import warnings
from sklearn.preprocessing import LabelEncoder
from sklearn.model_selection import StratifiedKFold,GroupKFold
from sklearn.decomposition import TruncatedSVD
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from tqdm import tqdm
import gc
import re
from sklearn.metrics import roc_auc_score
import os
from sklearn.feature_extraction.text import TfidfVectorizer, CountVectorizer
import polars as pl
from pathlib import Path
from glob import glob
import json
import joblib
import warnings
warnings.filterwarnings("ignore")

import os
os.makedirs('abstract')

import os
for i in range(106):
    os.system(f'python3 chatglm_abstract.py {i}')

t_dict = dict()
for i in tqdm(range(106)):
    a = joblib.load(f'abstract/t_dict_abstract_{i}.pkl')
    t_dict.update(a)

!rm -rf abastract

joblib.dump(t_dict,'t_dict_abstract_all.pkl')


train = pd.read_feather('data/train.feather')
valid = pd.read_feather('data/valid.feather')
test = pd.read_feather('data/test.feather')

piddf = joblib.load('data/pid_df.pkl')

data = pd.concat([train,valid,test]).reset_index(drop = True)
piddf = piddf[piddf['id'].isin(set(data['PID']))]


def cos_similarity(target, embedding):
    numerator = np.sum(target * embedding, axis=1)
    denominator = np.sqrt(np.sum(np.square(target)) * np.sum(np.square(embedding),axis=1))
    return numerator / denominator

ans = []
for autherID in tqdm(data['autherID'].unique()):
    pid_list = data[data['autherID'] == autherID]['PID'].to_list()
    for f1 in pid_list:
        x1 = t_dict.get(f1)
        x2 = [f for f in pid_list if f != f1]
        x2 = [t_dict.get(f) for f in x2]
        x1 = np.array(x1 , dtype = np.float32)
        x2 = np.array(x2 , dtype = np.float32)

        ans.append(list(cos_similarity(x1, x2)))

    
data['chatglm_pid_abstract_sim'] = ans

for f in ['chatglm_pid_abstract_sim']:
    data[f + '_mean'] = data[f].apply(lambda x:np.mean(x))
    data[f + '_max'] = data[f].apply(lambda x:np.max(x))
    data[f + '_min'] = data[f].apply(lambda x:np.min(x))
    data[f + '_std'] = data[f].apply(lambda x:np.std(x))
    data[f + '_median'] = data[f].apply(lambda x:np.median(x))
    
data = data.drop(['chatglm_pid_abstract_sim'],axis = 1)

data[['PID',  'autherID',  'chatglm_pid_abstract_sim_mean', 'chatglm_pid_abstract_sim_max',
       'chatglm_pid_abstract_sim_min', 'chatglm_pid_abstract_sim_std',
       'chatglm_pid_abstract_sim_median']].to_feather('chatglm_pid_abstract_sim.pkl')

t_dict = joblib.load(f't_dict_abstract_all.pkl')
abstract_emb = pd.DataFrame(t_dict.values()).values
svd_tmp = TruncatedSVD(n_components=16, n_iter=20, random_state=2024)
abstract_emb_svd = svd_tmp.fit_transform(abstract_emb)
abstract_emb_svd = pd.DataFrame(abstract_emb_svd,columns = [f'chatglm_abstract_emb_svd_{i}' for i in range(16)])
abstract_emb_svd = pd.concat([pd.DataFrame(t_dict.keys(),columns = ['PID']),abstract_emb_svd],axis = 1)
data = data.merge(abstract_emb_svd,on = ['PID'],how = 'left')

data[['PID', 'autherID', 'chatglm_abstract_emb_svd_0',
       'chatglm_abstract_emb_svd_1', 'chatglm_abstract_emb_svd_2',
       'chatglm_abstract_emb_svd_3', 'chatglm_abstract_emb_svd_4',
       'chatglm_abstract_emb_svd_5', 'chatglm_abstract_emb_svd_6',
       'chatglm_abstract_emb_svd_7', 'chatglm_abstract_emb_svd_8',
       'chatglm_abstract_emb_svd_9', 'chatglm_abstract_emb_svd_10',
       'chatglm_abstract_emb_svd_11', 'chatglm_abstract_emb_svd_12',
       'chatglm_abstract_emb_svd_13', 'chatglm_abstract_emb_svd_14',
       'chatglm_abstract_emb_svd_15']].to_feather('chatglm_pid_abstract_svd.pkl')

os.makedirs('title')

import os
for i in range(106):
    os.system(f'python3 chatglm_title.py {i}')

t_dict = dict()
for i in tqdm(range(106)):
    a = joblib.load(f'title/t_dict_title_{i}.pkl')
    t_dict.update(a)

!rm -rf title

joblib.dump(t_dict,'t_dict_title_all.pkl')


train = pd.read_feather('data/train.feather')
valid = pd.read_feather('data/valid.feather')
test = pd.read_feather('data/test.feather')

piddf = joblib.load('data/pid_df.pkl')

data = pd.concat([train,valid,test]).reset_index(drop = True)
piddf = piddf[piddf['id'].isin(set(data['PID']))]


def cos_similarity(target, embedding):
    numerator = np.sum(target * embedding, axis=1)
    denominator = np.sqrt(np.sum(np.square(target)) * np.sum(np.square(embedding),axis=1))
    return numerator / denominator

ans = []
for autherID in tqdm(data['autherID'].unique()):
    pid_list = data[data['autherID'] == autherID]['PID'].to_list()
    for f1 in pid_list:
        x1 = t_dict.get(f1)
        x2 = [f for f in pid_list if f != f1]
        x2 = [t_dict.get(f) for f in x2]
        x1 = np.array(x1 , dtype = np.float32)
        x2 = np.array(x2 , dtype = np.float32)

        ans.append(list(cos_similarity(x1, x2)))

    
data['chatglm_pid_title_sim'] = ans

for f in ['chatglm_pid_title_sim']:
    data[f + '_mean'] = data[f].apply(lambda x:np.mean(x))
    data[f + '_max'] = data[f].apply(lambda x:np.max(x))
    data[f + '_min'] = data[f].apply(lambda x:np.min(x))
    data[f + '_std'] = data[f].apply(lambda x:np.std(x))
    data[f + '_median'] = data[f].apply(lambda x:np.median(x))
    
data = data.drop(['chatglm_pid_title_sim'],axis = 1)

data[['PID',  'autherID',  'chatglm_pid_title_sim_mean', 'chatglm_pid_title_sim_max',
       'chatglm_pid_title_sim_min', 'chatglm_pid_title_sim_std',
       'chatglm_pid_title_sim_median']].to_feather('chatglm_pid_title_sim.pkl')

t_dict = joblib.load(f't_dict_title_all.pkl')
abstract_emb = pd.DataFrame(t_dict.values()).values
svd_tmp = TruncatedSVD(n_components=16, n_iter=20, random_state=2024)
abstract_emb_svd = svd_tmp.fit_transform(abstract_emb)
abstract_emb_svd = pd.DataFrame(abstract_emb_svd,columns = [f'chatglm_title_emb_svd_{i}' for i in range(16)])
abstract_emb_svd = pd.concat([pd.DataFrame(t_dict.keys(),columns = ['PID']),abstract_emb_svd],axis = 1)
data = data.merge(abstract_emb_svd,on = ['PID'],how = 'left')

data.columns

data[['PID', 'autherID','chatglm_title_emb_svd_0', 'chatglm_title_emb_svd_1',
       'chatglm_title_emb_svd_2', 'chatglm_title_emb_svd_3',
       'chatglm_title_emb_svd_4', 'chatglm_title_emb_svd_5',
       'chatglm_title_emb_svd_6', 'chatglm_title_emb_svd_7',
       'chatglm_title_emb_svd_8', 'chatglm_title_emb_svd_9',
       'chatglm_title_emb_svd_10', 'chatglm_title_emb_svd_11',
       'chatglm_title_emb_svd_12', 'chatglm_title_emb_svd_13',
       'chatglm_title_emb_svd_14', 'chatglm_title_emb_svd_15']].to_feather('chatglm_pid_title_svd.pkl')
